Extract normalize_url and normalize_title to concern
Closes #177
